+++ assets/l10n/ru-RU.ftl
@@ -435,7 +435,7 @@ label_introduction_description =
         • Вы не нажмёте кнопку "Выйти".
 
     Чтобы не потерять доступ к аккаунту, задайте пароль.
-label_kb = КБ
+label_kb = {$amount} КБ
 label_last_seen = Был(а)
 label_login = Логин
 label_login_hint = Ваш уникальный логин
